Us6 upgrade account

+++ app/controllers/wikis_controller.rb
@@ -1,4 +1,6 @@
 class WikisController < ApplicationController
+  rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
+
   def index
     @wikis = Wiki.all
   end
@@ -13,23 +15,24 @@ class WikisController < ApplicationController
 
   def update
     @wiki = Wiki.find(params[:id])
-
+    authorize @wiki
 
     if @wiki.update_attributes(wiki_params)
       redirect_to [@wiki.user, @wiki], notice: "Wiki updated"
     else
-      flash[:error] = "Unable to update wiki. Please try again."
+      flash[:alert] = "Unable to update wiki. Please try again."
       render :edit
     end
   end
 
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
     if @wiki.destroy
       redirect_to user_wikis_path, notice: "Wiki deleted"
     else
-      flash[:error] = "Unable to delete wiki. Try again."
+      flash[:alert] = "Unable to delete wiki. Try again."
       render :show
     end
   end
@@ -42,16 +45,21 @@ class WikisController < ApplicationController
   def create
     @user = User.find(params[:user_id])
     @wiki = @user.wikis.build(wiki_params)
-
+    authorize @wiki if @wiki.private == true
     if @wiki.save
       redirect_to [@user,@wiki] , notice: "Wiki created"
     else
-      flash[:error] = "There was a problem creating the wiki. Please try again."
+      flash[:alert] = "There was a problem creating the wiki. Please try again."
       render :new
     end
   end
 
   private
+
+  def user_not_authorized
+    flash[:alert] = "You are not authorized to perform this action"
+    redirect_to(request.referer || root_path)
+  end
 
   def wiki_params
     params.require(:wiki).permit(:title, :body, :private)
